backend/services/microservicio/requirements.txt
pymongo[srv]
pydantic[email]
uvicorn
fastapi
python-dotenv
httpx
motor
cloudinary
python-multipart
google-cloud-firestore
pyjwt